+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Release Date: 2019-??-?? Valhalla 3.0.3
 * **Bug Fix**
    * FIXED: Fixed a rare loop condition in route matcher (edge walking to match a trace).
+   * FIXED: Fixed VACUUM ANALYZE syntax issue.  [#1704](https://github.com/valhalla/valhalla/pull/1704)
 
 ## Release Date: 2019-01-14 Valhalla 3.0.2
 * **Bug Fix**